Simplify 'FilenameRule'. When a file contains a single toplevel class/interface then the file should be named same as that class/interface. In all other cases a descriptive PascalCase name should be used.
Modifications below were required to comply Ktlint code with change above:
* Extract class KtlintCommandLine from file "Main.kt" to separate class
* Rename files "main.kt" in Test resources to "Main.kt"
* Rename file "BaselineSupport" to "CurrentBaseline" as the class CurrentBaseline is the dominant result of the file
* Rename file "SuppressedRegionLocator" to "SuppressionLocatorBuilder". Functions had to be wrapped in an object as otherwise the file had to be named identical to the private data class SuppressionHint
