+++ ktlint/src/test/resources/cli/custom-ruleset/Main.kt
