+++ ktlint/src/test/resources/cli/no-code-style-error/Main.kt
